AI_Fellowship_LatinAmerica/apps/week10_deploying_ml_models_into_production/shoe_classification/requirements.txt
numpy
tensorflow
fastapi
streamlit
Pillow
uvicorn
python-multipart